task708-create-employment-interview and task726 fixed Valid and change fields type

Related work items: #705, #708, #726

+++ CardMicroService/src/main/java/com/bilgeadam/repository/entity/GraduationProject.java
@@ -15,20 +15,20 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class GraduationProject extends BaseEntity{
     @Id
     private String graduationId;
-    private int meetingAttendance;
-    private int teamworkCompatibility;
-    private int numberOfCompletedTasks;
-    private int interestLevel;
-    private int presentation;
-    private int retroScore;
-    private double averageScore;
+    private Double meetingAttendance;
+    private Double teamworkCompatibility;
+    private Double numberOfCompletedTasks;
+    private Double interestLevel;
+    private Double presentation;
+    private Double retroScore;
+    private Double averageScore;
     private String studentId;
     @Builder.Default
     private EStatus eStatus=EStatus.ACTIVE;
-    private int meetingAttendancePercentage ;
-    private int teamworkCompatibilityPercentage ;
-    private int numberOfCompletedTasksPercentage ;
-    private int interestLevelPercentage ;
-    private int presentationPercentage ;
-    private int retroScorePercentage ;
+    private Double meetingAttendancePercentage ;
+    private Double teamworkCompatibilityPercentage ;
+    private Double numberOfCompletedTasksPercentage ;
+    private Double interestLevelPercentage ;
+    private Double presentationPercentage ;
+    private Double retroScorePercentage ;
 }
